src/main/persistence/JsonWriter.java
package persistence;

import java.io.*;
import org.json.JSONObject;
import model.SongLibrary;

// Adapted from JsonSerializationDemo
// Represents a writer that writes JSON representation of songLibrary to file
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    //          be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of songLibrary to file
    public void write(SongLibrary sl) {
        JSONObject json = sl.toJson();
        saveToFile(json.toString(TAB));
    }

    private void saveToFile(String json) {
        writer.print(json);
    }

    public void close() {
        writer.close();
    }
}
